functions/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
arduino-iot-client==3.0.0
arduino_iot_cloud==1.4.0
asn1crypto==1.5.1
attrs==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.17.0
cloudevents==1.9.0
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.7.0
dotenv==0.9.9
firebase-admin==6.6.0
firebase-functions==0.1.2
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.57.0
frozenlist==1.5.0
functions-framework==3.8.2
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-logging==3.11.4
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-events==0.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
micropython-senml==0.1.1
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.30.0
packaging==24.2
pandas==2.2.3
pg8000==1.31.2
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruff==0.9.6
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
